Build on ARM
This ended up max'ing a double and a qreal.

+++ kdecoration/breezebutton.cpp
@@ -176,7 +176,7 @@ namespace Breeze
             QPen pen( foregroundColor );
             pen.setCapStyle( Qt::RoundCap );
             pen.setJoinStyle( Qt::MiterJoin );
-            pen.setWidthF( 1.1*qMax(1.0, 20/width ) );
+            pen.setWidthF( 1.1*qMax((qreal)1.0, 20/width ) );
 
             painter->setPen( pen );
             painter->setBrush( Qt::NoBrush );
